+++ package.json
@@ -20,7 +20,8 @@
         "mongoose": "^5.13.3",
         "reflect-metadata": "^0.1.13",
         "supertest": "^6.1.3",
-        "swagger-ui-express": "^4.1.6"
+        "swagger-ui-express": "^4.1.6",
+        "tsyringe": "^4.6.0"
     },
     "devDependencies": {
         "@types/express": "^4.17.13",
