+++ src/app.ts
@@ -10,6 +10,8 @@ import { errors } from 'celebrate';
 
 import {connect} from "./database";
 
+import "./container";
+
 import { AppError } from "./AppError";
 
 import swaggerFile from './swagger.json';
