+++ src/models/ICurrency.ts
@@ -1,0 +1,9 @@
+export interface ICurrency {
+    _id?: string;
+    name: string;
+    code: string;
+    valueInUSD: number;
+    created_at?: Date;
+    updated_at?: Date;
+
+}
