+++ tsconfig.json
@@ -56,8 +56,8 @@
         // "inlineSourceMap": true,                     /* Emit a single file with source maps instead of having a separate file. */
         // "inlineSources": true,                       /* Emit the source alongside the sourcemaps within a single file; requires '--inlineSourceMap' or '--sourceMap' to be set. */
         /* Experimental Options */
-        // "experimentalDecorators": true,              /* Enables experimental support for ES7 decorators. */
-        // "emitDecoratorMetadata": true,               /* Enables experimental support for emitting type metadata for decorators. */
+         "experimentalDecorators": true,              /* Enables experimental support for ES7 decorators. */
+         "emitDecoratorMetadata": true,               /* Enables experimental support for emitting type metadata for decorators. */
         /* Advanced Options */
         "skipLibCheck": true, /* Skip type checking of declaration files. */
         "forceConsistentCasingInFileNames": true, /* Disallow inconsistently-cased references to the same file. */
